Fixing issue with CartesianBlock.getPinCenterFlatToFlat
